+++ maintenance/rebuildElasticMissingDocuments.php
@@ -302,7 +302,7 @@ class rebuildElasticMissingDocuments extends \Maintenance {
 		);
 
 		$this->reportMessage( "\nInspecting documents ...\n" );
-		$cliMsgFormatter->setStartTime( microtime( true ) );
+		$cliMsgFormatter->setStartTime( (int) microtime( true ) );
 
 		foreach ( $rows as $row ) {
